requirements.txt
fastapi
uvicorn
python-multipart
pillow
replicate
requests